PartOf=mount-cryptosd-plain@%i.service does work as intended, but fails on bootup
The idea to automatically stop an instance of this unit when the corresponding mounting unit is stopped, works as intended.
Manually starting this unit or the corresponding mounting unit, also causes the other one to start, and stopping either one also stops the other one.
But for some reason this does not work on boot-up:
a. The udev rule starts an instance of this unlocking (opening) unit.
b. This unit unlocks (opens) a Cryptsetup device.
c. Another udev rule starts an corresponding instance of a mounting unit.
d. I suspect that that unit (started in c.) enters temporarily a "failed" state, maybe because `udisks2.service` is not ready yet.
e. What can be observed from the Systemd journal, is that Systemd stops the unlocking unit (from a.), which in turn also stops the corresponding mounting unit.

Omitting `PartOf=mount-cryptosd-plain@%i.service` "fixes" this.  Leaving it as a comment for now.
Reference: https://www.freedesktop.org/software/systemd/man/systemd.unit.html#PartOf=